Add the update for validations

+++ backend/habragasolina/src/main/java/bo/edu/usfa/gasolina/habragasolina/Service/GasStationService.java
@@ -5,6 +5,7 @@ import bo.edu.usfa.gasolina.habragasolina.Entities.GasStation;
 import bo.edu.usfa.gasolina.habragasolina.Entities.GasStationAvailability;
 import bo.edu.usfa.gasolina.habragasolina.Repository.AvailabilityRepository;
 import bo.edu.usfa.gasolina.habragasolina.Repository.GasStationRepository;
+import bo.edu.usfa.gasolina.habragasolina.Repository.UserRepository;
 import bo.edu.usfa.gasolina.habragasolina.Entities.GasStation;
 
 import java.util.ArrayList;
@@ -22,11 +23,11 @@ import java.util.Optional;
 public class GasStationService {
 
     private final GasStationRepository gasStationRepository;
-    private final AvailabilityRepository availabilityRepository;
+    private final UserRepository userRepository;
 
-    public GasStationService(GasStationRepository gasStationRepository, AvailabilityRepository availabilityRepository) {
+    public GasStationService(GasStationRepository gasStationRepository, UserRepository userRepository) {
         this.gasStationRepository = gasStationRepository;
-        this.availabilityRepository = availabilityRepository;
+        this.userRepository = userRepository;
     }
 
     public GasStation updateGasStation(Integer id, GasStation updatedGasStation) {
@@ -43,6 +44,7 @@ public class GasStationService {
     public GasStation saveGasStation(GasStation gasStation) {
         if(gasStation.getName().length() <= 3 ){
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Name must be longer than 3 characters");
+        // bad request es 400 para la validacion.
         }
         if(gasStation.getName().length() > 100){
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Name can't be longer than 100 characters");
@@ -54,12 +56,17 @@ public class GasStationService {
     }
 
     public void deleteGasStation(Integer id) {
+        if (userRepository.existsByGasStationId(id)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                    "You cannot delete the Gas Station with users");
+        }
         if (gasStationRepository.existsById(id)) {
             gasStationRepository.deleteById(id);
         } else {
             throw new RuntimeException("Gas station not found, id: " + id);
         }
     }
+
 
     public GasStation getGasStation(Integer id){
         Optional<GasStation> gasStation = gasStationRepository.findById(id);
